fix the int float problem

+++ app/ACM/Match.cc
@@ -1,6 +1,6 @@
 #include "match.h"
 
-Match::Match(Animal* _animal, Client* _client, int _score, QVector<QString> _rules) :
+Match::Match(Animal* _animal, Client* _client, float _score, QVector<QString> _rules) :
     animal(_animal), client(_client), score(_score), rules(_rules)
 {
